APR-006 inclir validações nas requisições

+++ src/main/java/com/atacado/presentes/api/controller/ProdutoController.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -46,6 +47,10 @@ public class ProdutoController {
 
         Optional<Produto> produto = produtosRepository.findById(id);
 
+        if (produto.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
         return ResponseEntity.ok().body(produto);
     }
 
@@ -69,9 +74,16 @@ public class ProdutoController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deletarProduto(@PathVariable("id") Long id) {
-
+        
+        Optional<Produto> produto = produtosRepository.findById(id);
+        
+        if (produto.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        
         produtosRepository.deleteById(id);
 
         return ResponseEntity.ok().body("Produto deletado com sucesso!");
+        
     }
 }
